Address PR comments
Revert splash screen back to its own activity (it doesn't need its own nav graph like I thought it might), and remove call to setGraph in MainActivity
