+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,10 @@
         <meta-data
             android:name="com.google.ar.core"
             android:value="required" />
-        <activity android:name="com.willowtree.vocable.MainActivity">
+
+        <activity
+            android:name=".splash.SplashActivity"
+            android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <action android:name="android.intent.action.VIEW" />
@@ -25,6 +28,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name="com.willowtree.vocable.MainActivity" />
         <activity android:name="com.willowtree.vocable.settings.SettingsActivity" />
     </application>
