+++ app/src/main/java/com/willowtree/vocable/splash/SplashActivity.kt
@@ -1,0 +1,29 @@
+package com.willowtree.vocable.splash
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import com.willowtree.vocable.BaseViewModelFactory
+import com.willowtree.vocable.MainActivity
+import com.willowtree.vocable.R
+
+class SplashActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val viewModel = ViewModelProviders.of(
+            this,
+            BaseViewModelFactory()
+        ).get(SplashViewModel::class.java)
+
+        viewModel.exitSplash.observe(this, Observer {
+            if (it) {
+                startActivity(Intent(this, MainActivity::class.java))
+                finish()
+            }
+        })
+    }
+}
