+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/splashFragment">
+    app:startDestination="@id/presetsFragment">
 
     <fragment
         android:id="@+id/presetsFragment"
@@ -17,20 +17,6 @@
             android:id="@+id/action_presetsFragment_to_settingsActivity"
             app:destination="@id/settingsActivity" />
     </fragment>
-    <fragment
-        android:id="@+id/splashFragment"
-        android:name="com.willowtree.vocable.splash.SplashFragment"
-        android:label="SplashFragment"
-        tools:layout="@layout/fragment_splash">
-        <action
-            android:id="@+id/action_splashFragment_to_presetsFragment"
-            app:destination="@id/presetsFragment"
-            app:popUpTo="@id/nav_graph"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <action
-        android:id="@+id/action_global_splashFragment"
-        app:destination="@id/splashFragment" />
     <fragment
         android:id="@+id/keyboardFragment"
         android:name="com.willowtree.vocable.keyboard.KeyboardFragment"
